feat: reduce dependencies + switch to flutter_inappwebview

+++ lib/src/_youtube/player_event_handler.dart
@@ -1,55 +1,31 @@
 import 'dart:async';
 import 'dart:convert';
-
 import 'package:flutter/foundation.dart';
 import 'package:omni_video_player/omni_video_player/models/omni_video_quality.dart';
 import 'package:omni_video_player/omni_video_player/models/video_player_callbacks.dart';
 import 'package:omni_video_player/omni_video_player/models/video_player_configuration.dart';
-import 'package:omni_video_player/src/_youtube/web/controller.dart';
+import 'package:omni_video_player/src/_youtube/controller.dart';
 import 'package:omni_video_player/src/_youtube/_model/youtube_player_state.dart';
-import 'package:webview_flutter/webview_flutter.dart';
 
-class YoutubeWebPlayerEventHandler {
-  YoutubeWebPlayerEventHandler(this.controller, this.options, this.callbacks) {
-    _events = {
-      'Ready': onReady,
-      'StateChange': onStateChange,
-      'PlaybackQualityChange': onPlaybackQualityChange,
-      'PlaybackRateChange': onPlaybackRateChange,
-      'PlayerError': onError,
-      'FullscreenButtonPressed': onFullscreenButtonPressed,
-      'VideoState': onVideoState,
-      'AutoplayBlocked': onAutoplayBlocked,
-    };
-  }
+class YoutubeMobilePlayerEventHandler {
+  YoutubeMobilePlayerEventHandler(
+    this.controller,
+    this.options,
+    this.callbacks,
+  );
 
-  final YoutubeWebPlaybackController controller;
+  final YoutubeMobilePlaybackController controller;
   final VideoPlayerConfiguration options;
   final VideoPlayerCallbacks callbacks;
 
   final Completer<void> _readyCompleter = Completer();
-  late final Map<String, ValueChanged<Object>> _events;
 
-  void call(JavaScriptMessage javaScriptMessage) {
-    if (controller.isDisposed) return;
-    final data = Map.from(jsonDecode(javaScriptMessage.message));
-    if (data['playerId'] != controller.playerId) return;
-
-    for (final entry in data.entries) {
-      if (entry.key == 'ApiChange') {
-        onApiChange(entry.value);
-      } else {
-        _events[entry.key]?.call(entry.value ?? Object());
-      }
-    }
-  }
-
-  Future<void> onReady(Object data) async {
+  Future<void> onReady([Object? data]) async {
     if (!_readyCompleter.isCompleted) _readyCompleter.complete();
   }
 
-  Future<void> onStateChange(Object data) async {
-    final stateCode = data as int;
+  Future<void> onStateChange(Object? data) async {
+    final stateCode = data is int ? data : int.tryParse(data.toString()) ?? -1;
 
     final playerState = YoutubePlayerState.values.firstWhere(
       (state) => state.code == stateCode,
@@ -62,11 +38,13 @@ class YoutubeWebPlayerEventHandler {
     if ((controller.duration == Duration(seconds: 1) ||
         controller.duration == Duration.zero)) {
       controller.isReady = false;
+      controller.pause(useGlobalController: false);
+      controller.hasStarted = false;
+
       final String duration = await controller.runWithResult("getDuration");
       final int? seconds = double.tryParse(duration)?.round();
 
       if (seconds != null && seconds > 0) {
-        // corrisponde a circa 11 giorni
         controller.isLive = seconds > 1000000 && kIsWeb;
         controller.duration = Duration(seconds: (seconds - 2));
       } else {
@@ -76,8 +54,17 @@ class YoutubeWebPlayerEventHandler {
       final config = options.videoSourceConfiguration;
 
       if (config.initialPosition.inSeconds >= 0) {
-        await controller.seekTo(config.initialPosition);
+        await controller.seekTo(
+          config.initialPosition,
+          skipHasPlaybackStarted: true,
+        );
         controller.hasStarted = false;
+      }
+
+      if (config.autoMuteOnStart) {
+        controller.mute();
+      } else {
+        controller.volume = config.initialVolume;
       }
 
       if (!config.autoPlay ||
@@ -87,20 +74,10 @@ class YoutubeWebPlayerEventHandler {
         controller.isPlaying = false;
       }
 
-      controller.run('unMute');
-
-      if (config.autoMuteOnStart) {
-        controller.mute();
-      } else if (controller.isMuted && !controller.hasStarted) {
-        controller.run('mute');
-      } else {
-        controller.volume = config.initialVolume;
-      }
-
       controller.playbackSpeed = config.initialPlaybackSpeed;
 
-      controller.isReady = true;
       callbacks.onControllerCreated?.call(controller);
+      controller.isReady = true;
     } else {
       if (playerState == YoutubePlayerState.playing &&
           controller.hasStarted == false) {
@@ -128,47 +105,52 @@ class YoutubeWebPlayerEventHandler {
     }
   }
 
-  void onPlaybackQualityChange(Object data) {
+  void onPlaybackQualityChange(Object? data) {
     if (data is String) {
       final quality = omniVideoQualityFromYouTube(data);
       controller.currentVideoQuality = quality;
     }
   }
 
-  void onPlaybackRateChange(Object data) {}
+  void onPlaybackRateChange(Object? data) {}
 
   void onApiChange(Object? data) {}
 
-  void onFullscreenButtonPressed(Object data) {}
+  void onFullscreenButtonPressed(Object? data) {}
 
-  void onError(Object data) {
-    options.globalKeyInitializer.currentState!.refresh();
+  void onError(Object? data) {
+    options.globalKeyInitializer.currentState?.refresh();
     controller.hasError = true;
   }
 
-  void onVideoState(Object data) {
-    final json = jsonDecode(data.toString());
+  void onVideoState(Object? data) {
+    if (data == null) return;
 
+    final json = jsonDecode(data.toString());
     final seconds = (json['currentTime'] ?? 0).truncate();
 
     if (seconds == 0) {
       return;
     }
+
     controller.currentPosition = Duration(seconds: seconds);
 
-    if (controller.currentPosition >= controller.duration) {
+    if (controller.currentPosition >= controller.duration &&
+        controller.duration != Duration(seconds: 1)) {
       controller.pause();
       controller.seekTo(Duration.zero);
     }
   }
 
-  void onAutoplayBlocked(Object data) {
+  void onAutoplayBlocked(Object? data) {
     debugPrint(
       'Autoplay was blocked by browser. '
-      'Most modern browser does not allow video with sound to autoplay. '
+      'Most modern browsers do not allow video with sound to autoplay. '
       'Try muting the video to autoplay.',
     );
   }
 
-  Future<void> get isReady => _readyCompleter.future;
+  Future<void> get isReady {
+    return _readyCompleter.future;
+  }
 }
